ares-proxy/src/main/java/com/workshop/ares/AresParser.java
package com.workshop.ares;


import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.core.JsonProcessingException;

import java.io.IOException;

public class AresParser {

    public static AresData parse(String jsonData) throws IOException {
        ObjectMapper objectMapper = new ObjectMapper();
        return objectMapper.readValue(jsonData, AresData.class);
    }
    
    public static String serialize(AresData aresData) throws JsonProcessingException {
        ObjectMapper objectMapper = new ObjectMapper();
        return objectMapper.writeValueAsString(aresData);
    }
}